refactor : 마이페이지 관련 로직 유저 도메인으로 이관
