+++ src/main/java/com/miracle/coordifit/user/controller/UserController.java
@@ -6,20 +6,25 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.Authentication;
 
 import com.miracle.coordifit.auth.service.IJwtService;
 import com.miracle.coordifit.common.dto.ApiResponseDto;
+import com.miracle.coordifit.user.dto.MyPageResponseDto;
 import com.miracle.coordifit.user.dto.ProfileUpdateRequestDto;
 import com.miracle.coordifit.user.model.User;
 import com.miracle.coordifit.user.service.IUserService;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RestController
 @RequestMapping("/api/user")
 @RequiredArgsConstructor
@@ -84,6 +89,24 @@ public class UserController {
 
 		} catch (Exception e) {
 			throw new RuntimeException("계정 활성화 실패: " + e.getMessage());
+		}
+	}
+
+	
+	@GetMapping("/{userId}/mypage")
+	public ResponseEntity<ApiResponseDto<MyPageResponseDto>> getMyPageInfo(
+		@PathVariable String userId,
+		Authentication authentication) {
+		try {
+			String currentUserId = authentication.getName();
+			MyPageResponseDto myPageInfo = userService.getMyPageInfo(userId, currentUserId);
+
+			log.info("마이페이지 정보 조회 완료: userId={}", userId);
+			return ResponseEntity.ok(ApiResponseDto.success("마이페이지 정보 조회 성공", myPageInfo));
+		} catch (Exception e) {
+			log.error("마이페이지 정보 조회 실패: userId={}", userId, e);
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+				.body(ApiResponseDto.error("마이페이지 정보 조회 실패: " + e.getMessage()));
 		}
 	}
 }
