--- src/main/java/com/miracle/coordifit/user/service/IMyPageService.java
@@ -1,8 +1,0 @@
-package com.miracle.coordifit.user.service;
-
-import com.miracle.coordifit.user.dto.MyPageResponseDto;
-
-public interface IMyPageService {
-
-	MyPageResponseDto getMyPageInfo(String userId, String currentUserId);
-}
