+++ src/main/java/com/miracle/coordifit/user/service/UserService.java
@@ -2,6 +2,7 @@ package com.miracle.coordifit.user.service;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.List;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -9,6 +10,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.miracle.coordifit.auth.dto.SignUpRequestDto;
 import com.miracle.coordifit.auth.service.IEmailService;
+import com.miracle.coordifit.post.dto.PostDto;
+import com.miracle.coordifit.post.repository.PostRepository;
+import com.miracle.coordifit.user.dto.MyPageResponseDto;
 import com.miracle.coordifit.user.dto.ProfileUpdateRequestDto;
 import com.miracle.coordifit.user.model.User;
 import com.miracle.coordifit.user.repository.UserRepository;
@@ -16,15 +20,13 @@ import com.miracle.coordifit.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-/**
- * 사용자 서비스 구현체
- */
 @Slf4j
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class UserService implements IUserService {
 	private final UserRepository userRepository;
+	private final PostRepository postRepository;
 	private final IEmailService emailService;
 	private final PasswordEncoder passwordEncoder;
 
@@ -32,28 +34,21 @@ public class UserService implements IUserService {
 	public User signUp(SignUpRequestDto signUpRequestDto) {
 		log.info("회원가입 시작: {}", signUpRequestDto.getEmail());
 
-		// 1. 이메일 중복 검사
 		if (!isEmailAvailable(signUpRequestDto.getEmail())) {
 			throw new IllegalArgumentException("이미 사용 중인 이메일입니다.");
 		}
 
-		// 2. 닉네임 중복 검사
 		if (!isNicknameAvailable(signUpRequestDto.getNickname())) {
 			throw new IllegalArgumentException("이미 사용 중인 닉네임입니다.");
 		}
 
-		// 3. 이메일 인증 코드 검증
 		if (!emailService.verifyCode(signUpRequestDto.getEmail(), signUpRequestDto.getVerificationCode())) {
 			throw new IllegalArgumentException("이메일 인증 코드가 올바르지 않습니다.");
 		}
 
-		// 4. 사용자 ID 생성
 		String userId = generateUserId();
-
-		// 5. 비밀번호 암호화
 		String password = passwordEncoder.encode(signUpRequestDto.getPassword());
 
-		// 6. 사용자 객체 생성
 		User user = User.builder()
 			.userId(userId)
 			.email(signUpRequestDto.getEmail())
@@ -61,10 +56,9 @@ public class UserService implements IUserService {
 			.nickname(signUpRequestDto.getNickname())
 			.loginTypeCode("A20001") // 이메일 로그인
 			.isActive("Y")
-			.createdBy(userId) // 자기 자신을 생성자로 설정
+			.createdBy(userId)
 			.build();
 
-		// 7. 데이터베이스에 저장
 		int result = userRepository.insertUser(user);
 		if (result <= 0) {
 			throw new RuntimeException("회원가입 처리 중 오류가 발생했습니다.");
@@ -100,14 +94,12 @@ public class UserService implements IUserService {
 		try {
 			User user = userRepository.selectUserByEmail(email);
 
-			// 암호화된 비밀번호 비교
 			if (user != null && passwordEncoder.matches(password, user.getPassword())) {
-				// 마지막 로그인 시간 업데이트
 				updateLastLoginTime(user.getUserId());
 				return user;
 			}
 
-			return null; // 인증 실패
+			return null;
 
 		} catch (Exception e) {
 			log.error("사용자 인증 중 오류 발생", e);
@@ -192,6 +184,21 @@ public class UserService implements IUserService {
 			log.error("계정 활성/비활성 중 오류 발생: userId={}", userId, e);
 			throw e;
 		}
+	}
+
+	@Override
+	@Transactional(readOnly = true)
+	public MyPageResponseDto getMyPageInfo(String userId, String currentUserId) {
+		MyPageResponseDto myPageInfo = userRepository.getMyPageInfo(userId);
+		if (myPageInfo == null) {
+			throw new RuntimeException("사용자 정보를 찾을 수 없습니다: " + userId);
+		}
+
+		List<PostDto> posts = postRepository.getUserPosts(userId);
+		myPageInfo.setPostsCount(posts.size());
+		myPageInfo.setPosts(posts);
+
+		return myPageInfo;
 	}
 
 	private String generateUserId() {
