--- src/main/resources/mapper/user/MyPageMapper.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.miracle.coordifit.user.repository.MyPageRepository">
-
-    <select id="getMyPageInfo" parameterType="String" resultType="com.miracle.coordifit.user.dto.MyPageResponseDto">
-        SELECT 
-            u.user_id AS userId,
-            u.nickname,
-            u.email,
-            f.s3_url AS profileImageUrl
-        FROM users u
-        LEFT OUTER JOIN file_info f ON u.file_id = f.file_id
-        WHERE u.user_id = #{userId} AND u.is_active = 'Y'
-    </select>
-
-    <select id="getUserPosts" parameterType="String" resultType="com.miracle.coordifit.post.dto.PostDto">
-        SELECT
-            p.post_id AS postId,
-            i.s3_url AS imageUrl
-        FROM posts p
-        INNER JOIN (
-            SELECT
-                pi.post_id,
-                f.s3_url,
-                ROW_NUMBER() OVER (PARTITION BY pi.post_id ORDER BY pi.created_at) AS rowrum
-            FROM post_images pi
-            INNER JOIN file_info f ON pi.file_id = f.file_id
-        ) i ON p.post_id = i.post_id AND i.rowrum = 1
-        WHERE p.user_id = #{userId} 
-        AND p.is_active = 'Y'
-        ORDER BY p.created_at DESC
-    </select>
-
-</mapper>
-
